Move AI summarization button to the calls transcription post. Use the generated transcription.
